+++ packages/collection/src/i18n.mjs
@@ -39,6 +39,7 @@ import { i18n as onyx } from '@freesewing/onyx'
 import { i18n as opal } from '@freesewing/opal'
 import { i18n as otis } from '@freesewing/otis'
 import { i18n as paco } from '@freesewing/paco'
+import { i18n as partlib } from '@freesewing/partlib'
 import { i18n as penelope } from '@freesewing/penelope'
 import { i18n as percy } from '@freesewing/percy'
 import { i18n as sabrina } from '@freesewing/sabrina'
@@ -49,6 +50,7 @@ import { i18n as shin } from '@freesewing/shin'
 import { i18n as simon } from '@freesewing/simon'
 import { i18n as simone } from '@freesewing/simone'
 import { i18n as skully } from '@freesewing/skully'
+import { i18n as sophie } from '@freesewing/sophie'
 import { i18n as sunny } from '@freesewing/sunny'
 import { i18n as sven } from '@freesewing/sven'
 import { i18n as tamiko } from '@freesewing/tamiko'
@@ -108,6 +110,7 @@ export const i18n = {
   opal,
   otis,
   paco,
+  partlib,
   penelope,
   percy,
   sabrina,
@@ -118,6 +121,7 @@ export const i18n = {
   simon,
   simone,
   skully,
+  sophie,
   sunny,
   sven,
   tamiko,