+++ packages/i18n/src/index.mjs
@@ -1,4 +1,6 @@
+import { designs } from './designs.mjs'
 import { flags } from './flags.mjs'
 import { measurements } from './measurements.mjs'
+import { optiongroups } from './optiongroups.mjs'
 
-export { flags, measurements }
+export { designs, flags, measurements, optiongroups }