+++ packages/react/components/LineDrawing/index.mjs
@@ -36,6 +36,8 @@ import { Octoplushy, OctoplushyFront /*, OctoplushyBack*/ } from './octoplushy.m
 import { Otis, OtisFront /*, OtisBack*/ } from './otis.mjs'
 import { Penelope, PenelopeFront /*, PenelopeBack*/ } from './penelope.mjs'
 import { Sabrina, SabrinaFront, SabrinaBack } from './sabrina.mjs'
+import { Sarah, SarahFront, SarahBack } from './sarah.mjs'
+import { Sophie, SophieFront, SophieBack } from './sophie.mjs'
 import { Simon, SimonFront, SimonBack } from './simon.mjs'
 import { Teagan, TeaganFront, TeaganBack } from './teagan.mjs'
 import { Tristan, TristanFront, TristanBack } from './tristan.mjs'
@@ -90,6 +92,8 @@ export const lineDrawingsFront = {
   otis: OtisFront,
   penelope: PenelopeFront,
 
+  sophie: SophieFront,
+  sarah: SarahFront,
   sabrina: SabrinaFront,
   simon: SimonFront,
 
@@ -135,6 +139,8 @@ export const lineDrawingsBack = {
   /*octoplushy: OctoplushyBack,*/
   /*otis: OtisBack,*/
   /*penelope: PenelopeBack,*/
+  sarah: SarahBack,
+  sophie: SophieBack,
   sabrina: SabrinaBack,
   simon: SimonBack,
   teagan: TeaganBack,
@@ -189,6 +195,8 @@ export const lineDrawings = {
   octoplushy: Octoplushy,
   otis: Otis,
   penelope: Penelope,
+  sarah: Sarah,
+  sophie: Sophie,
   sabrina: Sabrina,
   simon: Simon,
   teagan: Teagan,
@@ -340,6 +348,14 @@ export {
   Penelope,
   PenelopeFront,
   /*PenelopeBack,*/
+  // Sarah
+  Sarah,
+  SarahFront,
+  SarahBack,
+  // Sophie
+  Sophie,
+  SophieFront,
+  SophieBack,
   // Simon
   Simon,
   SimonFront,