+++ designs/hi/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change log for: @freesewing/hi
 
 
+## 4.4.0 (2025-11-06)
+
+### Fixed
+
+ - Double the size of the eye (#613)
+
 ## 3.0.0 (2023-09-30)
 
 ### Changed